+++ src/onegov.gazette/onegov/gazette/collections/__init__.py
@@ -1,8 +1,10 @@
 from onegov.gazette.collections.categories import CategoryCollection
 from onegov.gazette.collections.notices import GazetteNoticeCollection
+from onegov.gazette.collections.organizations import OrganizationCollection
 
 
 __all__ = [
     'CategoryCollection',
-    'GazetteNoticeCollection'
+    'GazetteNoticeCollection',
+    'OrganizationCollection',
 ]