+++ src/onegov.gazette/onegov/gazette/collections/organizations.py
@@ -1,16 +1,15 @@
 from onegov.core.orm.abstract import AdjacencyListCollection
-from onegov.gazette.models import Category
+from onegov.gazette.models import Organization
 
 
-class CategoryCollection(AdjacencyListCollection):
-    """ Manage a list of categories.
+class OrganizationCollection(AdjacencyListCollection):
+    """ Manage a list of organizations.
 
-    The list is ordered by the title, unless the ordering is set manually
-    (which should never occure in our case).
+    The list is ordered manually (through migration and/or backend).
 
     """
 
-    __listclass__ = Category
+    __listclass__ = Organization
 
     def get_unique_child_name(self, name, parent):
         """ Returns a unique name by treating the names as unique integers
@@ -19,7 +18,7 @@ class CategoryCollection(AdjacencyListCollection):
         """
 
         names = sorted([
-            int(result[0]) for result in self.session.query(Category.name)
+            int(result[0]) for result in self.session.query(Organization.name)
             if result[0].isdigit()
         ])
         next = (names[-1] + 1) if names else 1