+++ src/onegov.gazette/onegov/gazette/forms/__init__.py
@@ -2,6 +2,7 @@ from onegov.gazette.forms.category import CategoryForm
 from onegov.gazette.forms.common import EmptyForm
 from onegov.gazette.forms.common import RejectForm
 from onegov.gazette.forms.notice import NoticeForm
+from onegov.gazette.forms.organization import OrganizationForm
 from onegov.gazette.forms.user import UserForm
 
 
@@ -9,6 +10,7 @@ __all__ = [
     'CategoryForm',
     'EmptyForm',
     'NoticeForm',
+    'OrganizationForm',
     'RejectForm',
     'UserForm',
 ]