+++ src/onegov.gazette/onegov/gazette/layout.py
@@ -5,6 +5,7 @@ from onegov.core.static import StaticFile
 from onegov.gazette import _
 from onegov.gazette.collections import CategoryCollection
 from onegov.gazette.collections import GazetteNoticeCollection
+from onegov.gazette.collections import OrganizationCollection
 from onegov.gazette.models import Issue
 from onegov.gazette.models import Principal
 from onegov.user import Auth
@@ -79,6 +80,10 @@ class Layout(ChameleonLayout):
         return self.request.link(UserGroupCollection(self.app.session()))
 
     @cached_property
+    def manage_organizations_link(self):
+        return self.request.link(OrganizationCollection(self.app.session()))
+
+    @cached_property
     def manage_categories_link(self):
         return self.request.link(CategoryCollection(self.app.session()))
 
@@ -133,6 +138,11 @@ class Layout(ChameleonLayout):
             active = isinstance(self.model, UserGroupCollection)
             result.append((
                 _("Groups"), self.manage_user_groups_link, active
+            ))
+
+            active = isinstance(self.model, OrganizationCollection)
+            result.append((
+                _("Organizations"), self.manage_organizations_link, active
             ))
 
             active = isinstance(self.model, CategoryCollection)