+++ src/onegov.gazette/onegov/gazette/models/__init__.py
@@ -1,8 +1,9 @@
+from onegov.gazette.models.category import Category
 from onegov.gazette.models.notice import GazetteNotice
+from onegov.gazette.models.organization import Organization
 from onegov.gazette.models.principal import Issue
 from onegov.gazette.models.principal import IssueDates
 from onegov.gazette.models.principal import Principal
-from onegov.gazette.models.category import Category
 
 
 __all__ = [
@@ -10,5 +11,6 @@ __all__ = [
     'GazetteNotice',
     'Issue',
     'IssueDates',
+    'Organization',
     'Principal',
 ]